+++ src/pinkerton/main.py
@@ -19,10 +19,10 @@ def check_host(args) -> None:
         status_code: int = response.status_code
         body: str = response.text
 
-        status_error: str = f"[bold white on red][!] Host returned status code: {status_code} [/]"
+        status_error: str = f"[bold white on red][!] Host returned status code: {status_code}[/]"
 
         if response.ok:
-            print(f"[bold white on green]> Connected sucessfully with [bold white on yellow]{url}[/][/]")
+            print(f"[bold white on green][+] Connected sucessfully with [bold white on yellow]{url}[/][/]")
             extract_js(url, body)
         else:
             return print(status_error)
@@ -37,20 +37,20 @@ def extract_js(url, body) -> None:
     " Extract JavaScript files links from page source "
 
     # Connected sucessfully with target and start extractor
-    print(f"\n[bold white on green][*] Extracting JavaScript files from [white on yellow]{url}[/][/]")
+    print(f"[bold white on yellow][*] Extracting JavaScript files from [bold white]{url}[/][/]")
 
     re_jsfiles = r'src="(.*?\.js)(\?.*?)?"'
     jsfiles = re.findall(re_jsfiles, body)
 
     # Return number of JavaScript files found on the webpage source
-    print(f"[bold white on green][*] Scanning {len(jsfiles)} JavaScript file(s) [/]")
+    print(f"[bold white on green][+] Found {len(jsfiles)} JavaScript file(s) [/]")
 
     for jsfile, _ in jsfiles:
         final_url = f"{url}{jsfile}"
 
         if jsfile.startswith("http"):
-            print(f"[bold white on green] > Scanning: [bold white on yellow]{jsfile}[/][/]")
+            print(f"[bold white on green][+] Scanning:[bold white on yellow] {jsfile}[/][/]")
             direct_scan(jsfile)
         else:
-            print(f"[bold white on green] > Scanning: [bold white on yellow]{final_url}[/][/]")
+            print(f"[bold white on green][+] Scanning:[bold white on yellow] {final_url}[/][/]")
             passed_scan(final_url)